Move handler to utils and generalize
