+++ src/bots/ama/index.js
@@ -1,27 +1,9 @@
 import addEvent from './subCommands/addEvent';
 import client from '../../client';
+import { commandHandler } from '../../utils';
+import { getStrings } from './constants';
 import listEvents from './subCommands/listEvents';
-import { parseCommandString } from '../../utils/index';
-import { getCommandsString, getStrings } from './constants';
 
 const subCommands = { add: addEvent, list: listEvents };
 
-client.on('ama', () => {
-  let cmd = parseCommandString();
-  let subCommand = cmd.subCommand;
-
-  if (!subCommand) {
-    client.message.channel.send(getCommandsString());
-    return;
-  }
-
-  let targetCmd = subCommands[subCommand];
-
-  if (!targetCmd) {
-    client.message.channel.send(getStrings().invalidSubCommand);
-    client.message.channel.send(getCommandsString(subCommands));
-    return;
-  }
-
-  targetCmd.handler(cmd.arguments);
-});
+client.on('ama', () => commandHandler(subCommands, getStrings()));
